save state work

+++ src/com/tanqbay/targetshooter/Item.java
@@ -11,12 +11,12 @@ public class Item implements Comparable<Item>,Parcelable {
 	
 	protected boolean readyToBeRemoved = false;
 	protected int drawOrder = 0;
- protected double surfaceWidth;
-	protected double surfaceHeight;
+ protected float surfaceWidth;
+	protected float surfaceHeight;
 	
 	public Item(DrawingSurface drawingSurface){
-		surfaceWidth = drawingSurface.getWidth();
-		surfaceHeight = drawingSurface.getHeight();
+		surfaceWidth = (float) drawingSurface.getWidth();
+		surfaceHeight = (float) drawingSurface.getHeight();
 	}
 	
 	public void drawSelf(Canvas canvas){
